ath10k: add some flags to make memory allocations try harder.
GFP_DMA should have a qualifier of either GFP_KERNEL or GFP_ATOMIC.  With
neither, it neither waits nor has access to emergency memory, which means
maximizing its chance of disaster.

When allocating large memory areas, it can help to use __GFP_REPEAT, which
tries to free pages a few extra times.

And, because the current ath10k driver seems to be a little finicky about
memory allocation problems, let's just add __GFP_HIGH (to allow access to
the emergency pool) for all its allocations for now.  It's probably best to
remove that eventually, but should be okay in the short term since we're
also expanding the size of our emergency pool (in a separate patch).

Change-Id: I9b0933450c172e53a90218594bd6659fb8e7be4c