Make SSDP more independant of miniupnpd.
Split ssdpd into its own init script. It does not need to be restarted
when the FriendlyName or IP addresses changes, and its beneficial to
leave it running so it will retain the set of known services on the LAN.

Make /var/run more permissive so daemons won't have to run as root.

Build ssdp_poll in google_platform.

Change-Id: I4b5c48dbfc4309442954f33b2a1f8554a4ae4445