BUILDROOT_DL_NEVER didn't work.
When enabled, the error would trigger even for files already in
vendor/opensource/tarballs, because it didn't check for existence before it
checked for the variable.  So it was impossible to build anything.

This was counterbalanced by a bug in the autobuilder where it didn't
actually set BUILDROOT_DL_NEVER.

Change-Id: Ib9053c3db8f557f347f38c8211b513dfa9a37dae