S80upnpd: avoid use of 'eval', add babysitter, avoid arbitrary sleep delay.
'eval' in sh is extremely dangerous ("little bobby tables").  Let's use a
different trick instead to make it easier to print a command and then run
it.

miniupnpd wasn't running under a babysit instance, so if it ever died due to
a bug, it would accidentally stay dead.

Stopping/restarting upnpd assumed that it would shut down within 1 second,
which on a heavily loaded system, might not be the case.  Use killpid
instead, which finishes faster on average but does not return until the
process is actually dead.

Change-Id: Ic2a0b1c07c504887c616dcba262664753603379b