mwifiex: fix the 'passive' flag to iw scan.
Previously it would always actively scan, which caused the channel to be
spammed with probe requests if a lot of people do background scans.  This
makes it work as expected.

The offchannel time with active scans is 150-180ms as before.  Now passive
scans are more like 250ms, which is longer than I'd like.  With ath9k the
offchannel time is just slightly over 100ms, which is what I'd expect, so
this may still need some improvement, but at least it's better than it was.

Change-Id: Id0351be51bcf923156ae73108f247e4622871911