hostapd: add an alivemonitor.
The actual detection of liveness is a little tricky.  We look for any
interesting nl80211 frames sent by the kernel, which includes beacons (not
our own) and probes, as well as associate/disassociate, etc.  In a world
where there's at least one other AP on our channel sending beacons, that
should be a good indicator that things are working.

However, in a relatively idle environment, this could cause problems.
Luckily, with waveguide running, it'll occasionally generate background scan
events that also pass through nl80211 and ping the alivemonitor.

Change-Id: I2f742015a9f36b27f0665de75dd898ae7b0eb777