Make factory_status webserver visible from WAN.
If a USB key is present at boot containing a file named
"factory_status", we webserver is start on port 8883
containing information useful for the final check at the
factory.

We'd tested this from LAN ports, but the factory needs to
do this at OOBA (Out Of Box Audit) where there is no LAN
configuration. The WAN port is the only useable port.

If a user figures out to start a factory_status webserver
(hello to anyone reading this) nothing terrible happens.
The server is read-only, and nothing on the page is
sensitive.

Change-Id: I68a75724d863bd20564823599c0f7d7c4dcec177