/bin/cwmpd: kill run-isostream in case it was left running.
S85catawampus stops it when you restart cwmpd, but if cwmpd dies and the
babysitter restarts it, an isostream instance could have been left running
forever.  This would be a problem because the new cwmpd might start another
one but not be able to kill the original.

People running isostream from outside cwmpd are unlikely to use
run-isostream, so this won't get in the way of that.

b/17573901

Change-Id: I183a00ecb0e4ed22576336714bcb3cb071aa3b07