ath10k: call ath10k-reset in a couple more cases.
If, after a firmware crash, the driver was unable to read the fw dump, it
would just immediately return rather than trying a reset.  This was always
true, but also prevents ath10k-reset from getting called, so fix that.

Also, just for paranoia, add an extra echo to /dev/kmsg right at the top of
ath10k-reset that doesn't depend on logos or anything else.  There may be
odd cases where something goes wrong with the totally clear environment and
we'd like to know about them.

b/17404424

Change-Id: Ieacdc65d3dba7e0096d243db0936f8530e118276