package/sysvinit: Cross compile issue /usr/lib*
Fix Makefile to be more suitable for cross compilation. Previously, it
looked at /usr/lib*/libcrypt.a on the host system. This worked for
Ubuntu systems which had /usr/lib32/libcrypt.a which comes with the
Ubuntu package libc6-dev-i386. It turned out, though, that some systems
(including my corp laptop) don't have this file installed. Instead,
their libcrypto.a is in /usr/lib/x86_64-linux-gnu/libcrypto.a. That
said, the Makefile shouldn't look at /usr/lib in the first place.

Change-Id: I0521624a3308cfabf185f9c9d36860d4a2f2c73f