/bin/wifi: support primary channels other than the lowest in each group.
Previously, we would fail in -w80 or -w40 modes, with explicitly specified
channels, except where the specified channel was the very lowest channel in
the set.  So you could specify -c36 or -c149, but not -c40 or -c48.

This was intentional, since according to common wisdom, using eg. a 40 MHz
wide channel using 36+ (ie.  36+40 with 36 as the primary) will cause
interference problems if someone else is using 40- (ie. 36+40 with 40 as the
primary), so we were trying to limit the options to reduce conflicts.
However, according to our tests, the common wisdom is wrong, and mixing 36+
with 40- seems to have pretty much identical performance as putting
everybody on 36+, as long as all users are HT40.

But if some users are HT20 only, performance is better with the *mixed*
environment (36+ and 40- on different APs) since the two APs could be
talking to two different 20 MHz clients at once.  The same logic applies for
choosing the primary 40 MHz half of an 80 MHz channel (and the primary 20
MHz quarter inside that).

Note that this code only affects manual channel selection; the autochannel
algorithm doesn't yet attempt to choose unusual primary channels.

Note: an extra curveball is that hostapd sometimes ignores the primary
channel you request, and just swaps primary/secondary on your behalf.  That
behaviour doesn't seem to helpful for people who are trying to force the
channel number, but there's no hostapd config option to turn it off, so you
have to comment it out.

b/17291657

Change-Id: Ifd4631a39a79968da9fa04b47367d55b6820a8ca