Fix bug in recursive-repo handling of project downloading
If a sub-project was downloaded before its parent, the move handling for
the parent would fail (it's .tmp directory would end up as a child of the
target directory). For example, if 'foo/bar' was downloaded before 'foo/',
then the download for foo/ would end up creating 'foo/foo.tmp'.

This change fixes that issue by splitting the move into a mkdir and a move
of the project contents.

Note that since posix has no good way to move all files including hidden,
this approximation does miss files of the form '.?' (e.g. '.a'). But then
the rmdir will fail, so at least we'll spot it happening!

Change-Id: Ic7cab1e010f02473b63b42c408c31a06592118b3