Revert setup-dhclient from "Start dhclient ... for SpaceCast box."
This reverts the not-only-spacecast past of commit
1febe848b3bcea04b28457d11d26f2d1b0fa4f88.

I don't know what this was supposed to do, but it's a syntax error and
prevents dhclient from actually being configured on any platform.

All the extra comments are also not necessary, and the whitespace on
the newly added lines is very strange.

Change-Id: I4ec67dd5f8f7fc50d55b701e2314677e8bb02c87