{gfrg200,gftv100}_defconfig: enable avahi support in samba.
We originally turned this off so we wouldn't need dbus.  But now we have
dbus anyhow, so let's turn it back on.

With this, samba correctly registers itself with mDNS so Macbooks can see
it.  (Note: if you have a locked-down enterprise Macbook, it might still
ignore mDNS.)

Not enabled on gftv200 because that build never runs smbd.

Another option would be to provide an /etc/avahi/services/smb.service, but
some gftv100 devices (namely gfhd100) do not run samba, so that would cause
a misleading service advertisement.  Better to let smbd advertise itself
exactly when appropriate.

Change-Id: I8d97e39931f112e7ac148e350ce23788c8337785