S85catawampus and /etc/profile: set umask explicitly to 022.
We've had a few problems where people ssh into a system, restart catawampus,
and end up with files written with world-unreadable permissions so that
their target non-root processes can't actually read them.  It turns out that
ssh's default umask is 077.  This is hard to diagnose and ends up looking
like a catwampus bug.

Let's fix this in two places: every time we start a shell, and every time we
start cwmpd, for good measure.

Change-Id: I43c7354fc02e1084138e8031040fbc29c8a98bfb