coredump: use HDD instead of flash when available
This adds logic to the coredump script to dump cores into the HDD
instead of flash, and not overwrite any older core dumps. This will
be very helpful to collect more than just the most recent dump. E.g.,
in long-term soak tests sagesrv may crash several times and then
restart, but only the last core file is available.

To make things easier for the user a symbolic link (/user/last-core)
is generated to point to the latest core file in HDD.

Change-Id: I743099b1628955d8d53f58d6ea4319fd39938199