Make build work for rg200 and tv300.
The RG200 and SC100 configs build the simpleramfs but don't link it
directly into the kernel as an initramfs, instead it uses a split image
uImage format.  This was originally ifdef on zImage but now we have a
new platform (TV300) that uses zImage and wants to link in the
simpleramfs as an initramfs.

RG200 SC100 configs don't link as initramfs because those configs need
to include some kernel modules into the initramfs, so the kernel needs
to depend on simpleramfs, but simpleramfs needs to depend on the kernel.
Buildroot has some crufty ability to do this but it slows down making
simple changes, and rumor has it that it is buggy and error prone.

Change-Id: Iebef4781806398f6ac8428c4c92144424a2b4e54