spacecast: change rule to fix proto import path
The command to fix proto import path for SpaceCast only worked in the
proto/src/spacecast folder, leaving out protos generated for the
OnePlatform API.

This command uses the same command that is in the /vendor/google/spacecast
repo to fix the proto import path.

Also, added a comment explaining what does the command fix.

Change-Id: Ie333d850961a4e9629c61ed4d4d928ec095b8933