GFRG200: Move kernel load address to 0x04008000
The bootloader previously loaded the zImage at 0x03008000 which turned
out to be problematic if we append a device tree blob to the kernel. The
physical range from 0x02C00000 to 0x03FFFFFF is reserved for exclusive
use by PFE and is not accessible to Linux. The kernel would crash when
it tries to access the device tree blob that is located in this range.

Change-Id: I6f5625bf5af2ec73df6c63f3a783a4b8d6bbae06