Fix simpleramfs/target-finalize race condition
Previously, we had a race condition between the simpleramfs build cmds
and target-finalize because neither depended on the other:

target-finalize did not depend on simpleramfs because simpleramfs was
not a real package due to the fact that the simpleramfs Makefile was
located under fs/ instead of package/.
Likewise, simpleramfs did not depend on target-finalize because
simpleramfs is not a real fs as it uses the GENTARGETS macro instead of
ROOTFS_TARGET macro.

Running target-finalize and the simpleramfs build cmds is problematic
because both recipes potentially operate on the same files.

Change-Id: I61535cf309b050a3b44bd533feb81a9a2f859d0f