WindCharger: Changes disable/enable functions for wlan0/lan0
Now the lan and wlan devices should be properly disabled/enabled when
the temperature goes above/below the specified thresholds.

This is specifically related to use of the wifi daemon when disabling
wlan0 and then reenabling it.  Without using the daemon, wlan0 would be
shut down properly, but then it would not be possible to reenable the
device.

b/21331251

Change-Id: I89309d02f05b3ff8d79f3a0700ee44d18847a419