upload-logs: set a shorter connection timeout.
upload-logs tries IPv6 first, then IPv4.

The default pycurl connection timeout is 5 minutes,
but we set an alarm for 60 seconds at the top of upload-logs.

The result is:
+ systems with no IPv6 connectivity (no IPv6 address assigned)
    timeout IPv6 immediately and then try IPv4.
+ systems with broken IPv6 connectivity (IPv6 address is present
    but does not work) do not timeout IPv6 before the 60 second
    alarm fires.

This explains why many dogfooder systems only upload logs
once, at boot. They manage to upload between the time when
an IPv4 address is assigned and when a (nonfunctional) IPv6
address is assigned.

Change-Id: Ie2fee19e076058c99568d9bc829a3472109f1b9c