Add recoveryfs target
A recoveryfs consists of an initramfs image based on simpleramfs.gpio
containing a rootfs image.

NOTE: A recovery system allows console and potentially SSH access,
therefore the resulting kernel image must *not* be signed with the
production key. The bootloader must authenticate the recovery image
with a separate recovery key. On systems with a TPM the bootloader
needs to clear the TPM before booting the recovery image to prevent
access to sensitive data.

Change-Id: I1b4fe6cdda54c2f04548c25b8b47718959282328