spacecast: Allow SSH access with SpaceCast specific key
On recovery images SSH is enabled to allow login without access to the
serial console. The potential users of the recovery image like the
factory or repair facilities will have access to the private key, it is
assumed that the key can be leaked to the public. The private key is not
considered a secret since SSH is disabled on SpaceCast production builds.

For convenience SSH access with the "Bruno key" is still possible.

Change-Id: I124e945c69255be702ab850492d1e34b637b7544