Revert "Spacecast: load DVB driver later in the boot process to avoid conflicting with TPM."
This reverts commit 53f79005f04da8acfc0d8f2d8315c6ca54a846a5.

Rolling this back now that a proper fix is implemented. By putting this file back in its proper place,
it will help validate that the problem is really fixed and there is no timing race. Also, it is
possible that loading the tuner driver too late can cause other issues if it isn't loaded before tuner
configs are set.

Bug: 23689207
Change-Id: Ic4216b21655edcdfe3a52894094bca784c923ee9