emergency/buildroot: trigger cleanup script
buildroot configs that have the "HAVE_EXTRA_CLEANUP" option set will
trigger the cleanup script to create a trimmed down rootfs.

A side effect of selecting this new option will also cause the
TARGET_ROOTFS_RECOVERYFS option to be set as well. This will create
a slimmed down emergency image that is baked into the kernel small
enough to install to the reserved emergency partition.

Change-Id: If82f63789511e764da1367449045ccfbd47b372d