Switch tv200/254 to use lzo squashfs.
Our squashfs is currently compressed with LZMA which does give the
best compression ratio, but is terribly slow:
https://jonathancarter.org/2015/04/06/squashfs-performance-testing/

LZO is the fastest(**), it speeds up boot time by 15-20s
on tv200.

**LZ4 is faster than LZO but not in our current toolchain

Using a smaller block size, when doing page fetches we'd have to
read a lot less data from flash per fault but might end up having
more small reads, so it sort of depends on the paging pattern.

Space is limited on tv100, therefore not enabling on gftv100 yet.

Change-Id: I58665459318c920e7cfb6165e4489fb405b4d7b3