Create --unsigned flag to create unsigned builds
This change is designed to allow us to create production builds without
a dependency on Keystore, with the ultimate goal of automating release
builds.  With this change we can perform post-build signing which should
drastically reduce the amount of time required to create a release.

Some details can be found at http://go/gf-cpe-signing

Change-Id: If23750cdd16a73610ce1a8a8e8c3fe42e7b86e55