bluez: support sending unplug.
Install unplug-GFRM100 to /usr/bin, for the convenience
of cwmpd and Oregano (which will likely be running it).

Also, remove the last remnant of the BluezUnplugNoKeypress
experiment. That was an earlier, less sophisticated attempt
at working around GFRM100 misbehavior, but it did not work
well and was removed from the Blue/Z codebase.

Change-Id: I0c1d37417ead7451b4451546af90485e7c1257b2