Add sniproxy as a Buildroot package.
sniproxy will let the captive portal proxy HTTP and TLS to whitelisted
sites.  This is a prerequisite for Wireless TV in the current design.

sniproxy adds a couple new dependencies to the system:

* libev (builtin; updated to 4.22)
* pcre  (builtin)
* udns  (custom)

Update the builtin config for libev to 4.22; the provided download URL
returned a 404. The distribution site archives older versions and
recommends "So when you hardcode a URL, hardcode one from Attic/."

UDNS is required to support the hostname and hostname wildcard features
of sniproxy, which will be needed by the captive portal.

Change-Id: I3018b8418a8a18ff30293096d52b4bb9daffc23f