buildroot: refactor S09drivers back into one file
Refactor the /etc/init.d/S09drivers.* files back into a single file.
This is done in a similar way to the S40network.* refactor: Decide at
runtime which platform this is and apply the appropriate configuration.

Additionally add an is-ptp stanza for Chimera to rename the interface
eth0 to craft0 for the craft port.

Change-Id: Ifc65b7ae02ce8d8b6ae30a8ccf274d80d260bbf0