agetty: Disable parity detection
agetty tries to be smart and auto-detect the presence of a parity bit in
the RS-232 signal. We don't need this feature because everyone uses the
"8 data bits and no parity" setting. It sometimes happened to me that
agetty would misinterpret noise caused by hot plugging the RS-232
connection and switch to 7 bit communication which resulted in garbage
printed on the console.

Change-Id: I31bd5f9a2ca2f534fa2acc0f7a99a27b628cf128