Scrub only certain UBI volumes
Previously, S94ubiscrubber would launch a ubiscrubber for every UBI
volume it can find. The problem with this is that once ubiscrubber
opened a volume for reading, it cannot be ubidetached until the next
reboot. This turned out to be an issue when S94ubiscrubber kicked of a
ubiscrubber for rootfs-prep while ginstall was running.

b/27302048

Change-Id: I08d5513945e5144634dd78e53828f29406f3f563