Add a sniproxy config for autoprovisioning.
This is loosely based off of estruylov's cl/106187020, CPE-ified to
always run in the foreground.

Unlike Eugene, we're *only* trying to allow a TV Box to get to the ACS
long enough to download config. It doesn't need to be able to update
from versions before gfiber-47 (because those would need a wired upgrade
to be able to use wireless autoprovisioning anyway.)

Also, update /bin/connection_check to be compatible with a proxied ACS.

Change-Id: I0001108f267af2e29679e05b5bf2d5e042d7688b