quantenna: multiple interface support.
Previously, the qdpc-host driver created a single interface named wlanX
that handled all communication with the Quantenna device. This was
problematic for a number of reasons:

- For wireless provisioning and guest networks, multiple APs run with
  different security configurations. Traffic from each AP must be kept
  separate, which necessitates separate interfaces on the host.

- With cfg80211 drivers, the wlanX AP interface is in br0 while the
  wcliX client interface is not. With a single wlanX, the interface must
  be moved in and out of br0 when switching between AP and client modes.
  This can cause issues with forwarding and is inelegant.

- Code that can normally assume the existance of a wcliX client
  interface needs special cases for Quantenna.

This change introduces multiple interface support using VLAN interfaces
on top of the underlying qdpc-host interface (now called quantenna),
which addresses all of the above problems. On the Quantenna device, each
AP and client interface is configured as a VLAN access port with the
VLAN ID of the corresponding interface on the host, which separates
traffic between interfaces.

Change-Id: Icf8b00e0430217c5e2ae9371d5a6ea06dd9b7d4d