Give watchdog 7 seconds to handle SIGUSR1
Previously, we sent SIGTERM immediately after SIGUSR1 which did not
leave the watchdog script enough time to send the "V" character to the
kernel to stop the hardware watchdog. As a result, the hardware watchdog
was not stopped, and the watchdog rebooted the system after running the
poweroff command. The expected behavior is to not reboot.

This had been fixed by c7f8cb21c1d8f58503a2d973afd0dea0c1f305f5 but has
then subsequently been re-introduced by
344fb679587d2dcefeaeedd09877275750025fbc.

b/28826960
b/14420969

Change-Id: I3b8ca07cc35e8181d7c003802a81f39b21d9aa22