S40network: assign unique MAC address to Quantenna PCIe interface.
The Quantenna PCIe interface was assigned 02:00:00:00:00:00 for all
devices, with the requirement that no packets originate from the
interface itself, but from a bridge containing the interface instead.
This is not the case on GFTV254, where the interface is removed from the
bridge in client mode, resulting in a conflict when associated to
another Quantenna device like GFRG250.

Change-Id: I38d06b8120c945a097486cb6884498e02d11063e