psstat: extra_ping_servers and fix some platforms.
Provide a way to extend the list of sites to ping.
catawampus can populate /tmp/extra_ping_servers
with a comma-separated list of sites to ping once per
hour.

Also:
The attempt to disable psstat on Spacecast resulted in
also disabling it on GFMS100 and GFCH100.

Make psstat to run by default, exclude it on specific
platforms.

Change-Id: Ibc454dde69605f23e8ae3911871f65764b8b5237