Enable s6-rc on gfiber devices, on top of existing init scripts.
The s6 service manager can be used either as an additional service manager
(on top of init) or standalone (replacing init).  Eventually we want to do
the latter, but for the first round, let's do the former, and convert only a
relatively harmless service (isoping) to the new system.

Advantages of switching to the new system:
- based on a djb design (daemontools), so clean, fast, and secure.
- no more stupid pkillwait (and associated delays / slowness / fishiness).
- integrated locking, so when we want to restart a service, we won't end up
  actually running it more than once.
- most services can be started without a sh script, saving a lot of forking.
- services can be started in parallel, which should (once we get things
  converted) trim a lot off boot time.
- logos now also gets a babysitter.
- fewer copies of sh left running in the process tree.

I also looked at daemontools (djb), procd (openwrt), upstart, and systemd.
All of them share the characteristic of being much faster and less hacky
than init scripts but only s6 does everything we want without creating a
mess (eg.  it's maintained, it easily supports logos, and there's no dbus
dependency).

Change-Id: I85e344b8e0ff060efebb379e598e10c4952c30bd