spacecast: Add script to reset the tuner
Sometimes the USB bridge doesn't work properly after bringing the DVB
network interface down, probably due to a timing issue in the firmware
of the bridge. Recovery is possible through a hard reset of the USB
bridge. The reset script can be called from the appliance before it
sets up the DVB network interface.

Change-Id: I520402627889f7cff77f9fd16933391e8968f620
Google-Bug-Id: 25378377