repo:// urls: fix branch name copying on checkout.
Previously, when copying a git repo into the out/build/ directory, the copy
would retain the same branch name as checked out in the original, so when
you committed files in either repo, they'd continue to be visible in the
original.  The switch to git-worktree in
Ia369809f3bc5f4af221831eb81929d3bee957949 accidentally broke that, by always
checking out a detached HEAD.  Bring back the old behaviour.

Change-Id: I218b7637bea7f812d61a1815254acd95337b8c8e