buildroot: make lasertune be less verbose
Before we receive a PON-ID PLOAM message, we are polling aggressively
(once per second) and each poll generates a log message. If no PON-ID
PLOAM message is received (e.g. fiber is not connected) then this can
go on for an extended period of time generating a lot of noise.

Change-Id: I95baf39b974ec6bdd57dda6a4fa83728ae1f7487