google_dashplayer: Use the install targets for INSTALL commands
This shifts the responsibility for the install from buildroot to the
vendor/google/dashplayer code. This is desirable because it means that
any change to installed headers/libs/etc doesn't have to be coordinated
across two respositories. It also brings the INSTALL commands in line
with BUILD, HOST_TEST, etc.

Change-Id: I8f97261201ea25702b36511c1e12f66d7a915f00