S40wait-network: don't wait for all of svscan to finish.
It turns out the s6-rc command holds a global lock, so that running
  s6-rc -u change network firewall
while we are already running
  s6-rc -u change all
will actually wait until *all* the svscan-run processes have finished, not
just the ones (network and firewall) you requested.

The particular script which revealed the problem is the 'wifi' startup
script on GFHD254, which sits around for a really long time.  There's no
reason TV needs to be delayed by that.

Change-Id: Iafe1d812e4e33e92b51e0f2356dc294169df9cbb