ipapply:  Apply netmasks to IP addresses.
This reverts part of go/fibercl/77060 (buildroot: c3075d2), which
applied /32 netmasks to all IP addresses to prevent the kernel
automatically adding routes.  We wanted to prevent that because having
multiple routes (e.g. one wired and one wifi) with the same metric
breaks things, and there was no way to give the kernel a metric to
apply to the routes it created.

But that broke Cast, which is probably worse than the race condition
that results from deleting the routes immediately after the kernel
adds them.

Change-Id: Ied2e8e93388e8445f992e5ca391698482ad7a858