psstat: traceroute the extrapingservers
If we have a set of ExtraPingServers to monitor also run
traceroute to them, one site per hour.

The output format is on one line:
traceroute: google.com 96.120.89.53(13.054ms) 68.85.100.221(11.012ms) \
  162.151.78.9(11.112ms) 68.86.90.93(11.243ms) 68.86.86.26(11.867ms) \
  23.30.206.102(11.110ms) 108.170.242.241(13.309ms) 108.170.235.239(11.197ms) \
  216.58.195.78(11.531ms)

Only doing one site at a time is important: routers implement rate
limiting for the ICMP messages which traceroute depends on. If we send
a bunch of traceroutes all at once, the first few routers in the path
(which are the same for all destinations) will stop sending ICMP messages.
Spacing the traceroutes out to be one at a time gets much more reliable
results.

Also add environment variables to allow the extra_ping_servers
functions to be tested, and a psstat_network_box_test
script to be used in a gftest for the extra_ping_servers
and gstatic functions.

Change-Id: I0df70644372b5d1820e00df706e2c49fffb07a2e