Don't strip kernel symbols needed for stacktraces
Buildroot currently uses strip --strip-unneeded on all .ko files it
finds. It turned out that this command strips symbols that are necessary
to generate correct back traces (dump_stack()). Today, we are getting
back traces with incorrect symbol names.

To analyse the increase in size, I calculated the sum of the file sizes
of all .ko files in the respective target directories. All numbers in
bytes:
GFRG200 from 4352452 to 4828156 (+11%)
GFTV100 from 9786208 to 9910332 (+1%)
GFTV200 from 2885280 to 3127196 (+8)
GFTV254 from 1484840 to 1645702 (+11%)

For example, on GFRG200 we are looking at an increase of 464kB before
the image is compressed.

Change-Id: I23a5099278f5010edecb60eb497b91ee251b6b26