gfrg240: update defconfig
include BR2_PACKAGE_GOOGLE_TEST=y
We need a number of the utilities

use BR2_PACKAGE_PROTOBUF_ONLY_LITE
gfrg240 does not run ads code, the main use of
libproto is statscatcher. We can get by with just
the lite version of the protobuf library, which
reduces the image size a bit.

Change-Id: I697d30b28ff8f9d29e73b7a40d35c0fff7652992