Add patch for ChaCha20-Poly1305 support (second try)
This is based on the Cloudflare patch from
https://github.com/cloudflare/sslconfig/tree/master/patches
but with all code from boringssl updated to its latest version
and the main evp rewritten to work with it. The draft version
support was removed and the assembly code wasn't ported over
so this only uses the C version of the implementation.

Using very large blocks the performance is around 21MB/s on CPE1/2
and on RG210 the performance is slightly faster by 5% than AES-GCM.
A port including optimized code would be even faster on ARM.

Change-Id: Ic3d57280fcb93c763b8cbe1b7a2370ed9a16013e