Make host-libpcap dependency for waveguide only.
gflt110 build was failing on my workstation when building
host-libpcap, due to unsatisfied libnl (<netlink/>) dependency. libnl
in turn wouldn't build because of unsatisfied glib2 dependency. glib2
wouldn't build at all on the gflt110.

Turns out host-libpcap is only needed for waveguide. Waveguide is not
built on gflt110.

So this CL makes it so host-libpcap is built only if waveguide is
getting built.

Change-Id: If336f0215ca75e3a3bd78030d0eefa768bee60d3