Remove remains of failed merge attempt (initramfs=true)
The line previous to initramfs=true is missing a backslash character
("\") at the end. As a result, initramfs=true is not part of
LINUX_MAKE_FLAGS. This appears to be the result of a failed merge
attempt from over four years ago.

The Linux kernel apparently builds just fine even without initramfs=true,
so let's just remove it.

Change-Id: I23016b107f45fc589975fec03bdecca9268eda85